Time: 3 ms (7.97%) | Memory: 10 MB (5.87%) - LeetSync

+++ 118-pascals-triangle/pascals-triangle.cpp
@@ -2,13 +2,13 @@ class Solution {
 public:
     vector<vector<int>> generate(int numRows) {
         vector<vector<int>> res={{1}};
-        for(size_t i = 0 ; i < numRows-1; i++){
+        for(int i = 0 ; i < numRows-1; i++){
             vector<int> temp = {0};
             temp.insert(temp.end(),res.back().begin(),res.back().end());
             temp.push_back(0);
             vector<int> row;
 
-            for(size_t j = 0; j < res.back().size()+1;j++){
+            for(int j = 0; j < res.back().size()+1;j++){
                 row.push_back(temp[j]+temp[j+1]);
             }
             res.push_back(row);
